Added electric system that simulates all electric busses and can be connected to ground power Added manually exteneded ram air turbine to EFM (only used by electric system for now, not yet used in hydraulic system) Changed Radar, Terrain, RWR, WeaponSystem and Radio to now be powered by the correct bus Added canopy system to open/close canopy Fixed some warnings in the EFM code Fixed roughmet skin for 2 F-104G skins Changed debug console to be enabled when EFM dll is compiled as "Debug" Changed encoding from unicode to ascii Removed not supported x86 build target

+++ ExternalFM/ED_FM_Template/Table.h
@@ -23,7 +23,7 @@ public:
 		m_max(max)
 	{
 		m_dx = (max - min) / ((double)table.size()-1);
-		m_minIndex = min / m_dx;
+		m_minIndex = (int)(min / m_dx);
 	}
 
 	double operator()(double value)
@@ -34,8 +34,8 @@ public:
 
 		double index = (value - m_min) / m_dx;
 
-		int lower = floor(index);
-		int upper = ceil(index);
+		int lower = (int)floor(index);
+		int upper = (int)ceil(index);
 
 		if (lower == upper)
 			upper++;
@@ -68,7 +68,7 @@ public:
 		m_max( max )
 	{
 		m_dx = (max - min) / ((double)table.size() - 1);
-		m_minIndex = min / m_dx;
+		m_minIndex = (int)(min / m_dx);
 	}
 
 	double operator()( double value )
@@ -79,8 +79,8 @@ public:
 
 		double index = (value - m_min) / m_dx;
 
-		int lower = floor( index );
-		int upper = ceil( index );
+		int lower = (int)floor( index );
+		int upper = (int)ceil( index );
 
 		if ( lower == upper )
 			upper++;
